modularized tests, renamed submitButton in htmls for consistency, adjusted tests as needed
