+++ cms-force/e2e/cfg/tsconfig.e2e.json
@@ -1,7 +1,7 @@
 {
-  "extends": "../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "../out-tsc/app",
+    "outDir": "../../out-tsc/app",
     "module": "commonjs",
     "target": "es5",
     "types": [
