+++ cms-force/e2e/run_all_tests.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+# Runs protractor tests
+
+protractor cfg/test.conf.js
+read -n 1 -s -r -p "Press any key to continue"
