+++ cms-force/e2e/run_content_creator_test.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+# Runs protractor tests
+
+protractor cfg/ccreator.conf.js
+read -n 1 -s -r -p "Press any key to continue"
