+++ cms-force/e2e/run_module_creator_test.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 # Runs protractor tests
 
-protractor test.conf.js
+protractor cfg/mcreator.conf.js
 read -n 1 -s -r -p "Press any key to continue"
