+++ cms-force/e2e/run_population.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 # Runs protractor population script
 
-protractor populate.conf.js
+protractor cfg/populate.conf.js
 read -n 1 -s -r -p "Press any key to continue"
