+++ cms-force/e2e/src/test/contentcreation.po.ts
@@ -155,7 +155,7 @@ export class AppPage {
 
     private getSubmitButton(){
       // return element(by.xpath('/html/body/app-component/div/app-create-user/form/button'));
-      return element(by.id("submitbutton"));
+      return element(by.id("submitButton"));
     }
 
     clickSubmitButton(){
