+++ cms-force/e2e/src/test/contentsearch.e2e-spec.ts
@@ -1,4 +1,4 @@
-import { SearchPage } from './search.po';
+import { SearchPage } from './contentsearch.po';
 import { browser, logging } from 'protractor';
 
 describe('workspace-project App', () => {
