+++ cms-force/package.json
@@ -20,6 +20,8 @@
       "@angular/platform-browser-dynamic": "~7.2.0",
       "@angular/router": "~7.2.0",
       "@ng-select/ng-select": "^2.20.0",
+      "@types/jasmine": "^3.3.13",
+      "@types/jasminewd2": "^2.0.6",
       "rxjs": "~6.3.3",
       "zone.js": "~0.8.26"
    },
