+++ cms-force/src/app/components/content-creator-page/content-creator-page.component.html
@@ -16,7 +16,7 @@
                <div class="col-1"></div>
 
                <div class="col">
-                  <input id="titleTextBox"[(ngModel)]="title"
+                  <input id="titleTextBox" [(ngModel)]="title"
                      required
                      name="title"
                      type="text"
@@ -80,12 +80,13 @@
                      <div class="form-check form-check-inline"
                         *ngFor="let format of formats">
 
-                        <input  [(ngModel)]="selFormat"
+                        <input [(ngModel)]="selFormat"
                            class="form-check-input"
                            type="radio" name="formatOptions"
                            id="{{format}}" value="{{format}}">
                         <label class="form-check-label"
-                           for="{{format}}" id= "{{format}}RadioButton">{{format}}</label>
+                           for="{{format}}"
+                           id="{{format}}RadioButton">{{format}}</label>
                      </div>
                   </div>
                </div>
